refactor(poo): refactor in the content of reservation class.
